+++ src/Domain/Repository/ChatSessionRepositoryInterface.php
@@ -24,4 +24,6 @@ interface ChatSessionRepositoryInterface
     public function findOneById(int $id): ?ChatSession;
 
     public function hasMessageFromUser (ChatSession $chatSession, User $user): bool;
+
+    public function countCallVisioByEvent(Event $event):int;
 }