+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- VIMEET-2268 - Cal visio analytics on admin dashboard
+
 ## [2.68.0] - 2020-11-09
 ### Added
 - VIMEET-2198 - Broadcast webinars in HLS